+++ src/components/NavBar/Dropdown/ListMenu/ListTemplate/menuTemplates.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function MenuTemplates() {
+    return (
+        
+    )
+}
+
+export default MenuTemplates
