+++ src/components/NavBar/Dropdown/ListMenu/stateChange.jsx
@@ -1,0 +1,15 @@
+import { useState } from "react"
+
+function stateChange() {
+  const [onMouse, setOnMouse] = useState(false)
+
+  const handleOnMouse = () => {
+    setOnMouse(!onMouse)
+  }
+  return {
+    onMouse,
+    handleOnMouse,
+  }
+}
+
+export default stateChange
